Add docs. Slight refactor.  Add pbmc full example.
